Extract lsir.copy expansion into PostRegAllocLegalization pass.
Moves copy-to-mov expansion and redundant-mov elimination out of
RegisterColoring into a new PostRegAllocLegalization pass, run after
coloring in the reg-alloc pipeline. In PostRegAllocLegalization,
MovInstOpPattern (benefit=2) eliminates self-copies before
CopyExpansionPattern (benefit=1) expands the remaining copies to
hardware mov instructions.
