+++ include/aster/Interfaces/InstOpInterfaces.h
@@ -19,5 +19,13 @@
 #include "aster/Interfaces/InstOpInterface.h"
 
 #include "aster/Interfaces/InstOpInterfaces.h.inc"
+#include "mlir/IR/PatternMatch.h"
+
+namespace mlir::aster::detail {
+/// Canonicalization pattern for MovInstOpInterface, if the source and
+/// destination are the same register, the instruction can be removed.
+LogicalResult canonicalizeMovInstImpl(MovInstOpInterface mov,
+                                      RewriterBase &rewriter);
+} // namespace mlir::aster::detail
 
 #endif // ASTER_INTERFACES_INSTOPINTERFACES_H
