+++ lib/Dialect/AMDGCN/Transforms/CMakeLists.txt
@@ -15,6 +15,7 @@ add_mlir_library(AMDGCNTransforms
   OptimizeAMDGCN.cpp
   OptimizeInterferenceGraph.cpp
   Pipelines.cpp
+  PostRegAllocLegalization.cpp
   PreColoringLegalization.cpp
   PreloadLibrary.cpp
   RegisterColoring.cpp
