+++ lib/Dialect/AMDGCN/Transforms/RegisterColoring.cpp
@@ -16,8 +16,6 @@
 #include "aster/Dialect/AMDGCN/Transforms/Transforms.h"
 #include "aster/Dialect/LSIR/IR/LSIROps.h"
 #include "mlir/Analysis/DataFlowFramework.h"
-#include "mlir/IR/Builders.h"
-#include "mlir/Interfaces/FunctionInterfaces.h"
 #include "mlir/Transforms/GreedyPatternRewriteDriver.h"
 #include "llvm/ADT/EquivalenceClasses.h"
 #include "llvm/ADT/STLExtras.h"
@@ -160,20 +158,6 @@ struct RegInterferenceOpPattern : public OpRewritePattern<RegInterferenceOp> {
   using Base::Base;
 
   LogicalResult matchAndRewrite(RegInterferenceOp op,
-                                PatternRewriter &rewriter) const override;
-};
-
-struct CopyOpPattern : public OpRewritePattern<lsir::CopyOp> {
-  using Base::Base;
-
-  LogicalResult matchAndRewrite(lsir::CopyOp op,
-                                PatternRewriter &rewriter) const override;
-};
-
-struct MovInstOpPattern : public OpInterfaceRewritePattern<MovInstOpInterface> {
-  using Base::Base;
-
-  LogicalResult matchAndRewrite(MovInstOpInterface op,
                                 PatternRewriter &rewriter) const override;
 };
 } // namespace
@@ -501,111 +485,6 @@ RegInterferenceOpPattern::matchAndRewrite(RegInterferenceOp op,
   return success();
 }
 
-LogicalResult CopyOpPattern::matchAndRewrite(lsir::CopyOp op,
-                                             PatternRewriter &rewriter) const {
-  // Try to canonicalize the operation.
-  if (succeeded(op.canonicalize(op, rewriter)))
-    return success();
-
-  // If the result is used, bail out.
-  if (op.getTargetRes() && !op.getTargetRes().use_empty())
-    return failure();
-
-  // Get the source allocas. Bail out if the allocas are missing or need
-  // allocation.
-  FailureOr<ValueRange> srcAlloc = getAllocasOrFailure(op.getSource());
-  if (failed(srcAlloc) ||
-      llvm::any_of(*srcAlloc, [](Value v) { return needsAllocation(v); }))
-    return failure();
-
-  // Get the target allocas. Bail out if the allocas are missing or need
-  // allocation.
-  FailureOr<ValueRange> tgtAlloc = getAllocasOrFailure(op.getTarget());
-  if (failed(tgtAlloc) ||
-      llvm::any_of(*tgtAlloc, [](Value v) { return needsAllocation(v); }))
-    return failure();
-
-  assert(srcAlloc->size() == tgtAlloc->size() &&
-         "source and target allocas must have the same size");
-
-  assert(srcAlloc->size() > 0 &&
-         "source and target allocas must have at least one alloca");
-
-  auto srcTy = dyn_cast<AMDGCNRegisterTypeInterface>(op.getSource().getType());
-  auto tgtTy = dyn_cast<AMDGCNRegisterTypeInterface>(op.getTarget().getType());
-
-  // Bail if the source or target is not an AMDGCN register type.
-  if (!srcTy || !tgtTy)
-    return failure();
-
-  // Bail if the copy cannot be performed.
-  if (srcTy.getRegisterKind() != RegisterKind::SGPR &&
-      tgtTy.getRegisterKind() == RegisterKind::SGPR) {
-    return rewriter.notifyMatchFailure(
-        op, "cannot copy between non-sgpr type to an sgpr type");
-  }
-  if (!llvm::is_contained(
-          {RegisterKind::SGPR, RegisterKind::VGPR, RegisterKind::AGPR},
-          srcTy.getRegisterKind()) ||
-      !llvm::is_contained(
-          {RegisterKind::SGPR, RegisterKind::VGPR, RegisterKind::AGPR},
-          tgtTy.getRegisterKind())) {
-    return rewriter.notifyMatchFailure(
-        op, "cannot copy if data type is not SGPR, VGPR, or AGPR");
-  }
-  // AGPR copies: only AGPR->AGPR is supported (via v_accvgpr_write_b32).
-  if (tgtTy.getRegisterKind() == RegisterKind::AGPR &&
-      srcTy.getRegisterKind() != RegisterKind::AGPR) {
-    return rewriter.notifyMatchFailure(
-        op,
-        "cannot copy non-AGPR to AGPR (use v_accvgpr_write_b32 explicitly)");
-  }
-  if (srcTy.getRegisterKind() == RegisterKind::AGPR &&
-      tgtTy.getRegisterKind() != RegisterKind::AGPR) {
-    return rewriter.notifyMatchFailure(
-        op, "cannot copy AGPR to non-AGPR (use v_accvgpr_read_b32 explicitly)");
-  }
-
-  auto copyReg = [&](Value src, Value tgt) {
-    if (tgtTy.getRegisterKind() == RegisterKind::SGPR) {
-      SMovB32::create(rewriter, tgt.getLoc(), tgt, src);
-      return;
-    }
-    if (tgtTy.getRegisterKind() == RegisterKind::AGPR) {
-      VAccvgprMovB32::create(rewriter, tgt.getLoc(), tgt, src);
-      return;
-    }
-    VMovB32::create(rewriter, tgt.getLoc(), tgt, src);
-  };
-
-  // Create the copy operations.
-  for (auto [src, tgt] : llvm::zip_equal(*srcAlloc, *tgtAlloc))
-    copyReg(src, tgt);
-  rewriter.eraseOp(op);
-  return success();
-}
-
-LogicalResult
-MovInstOpPattern::matchAndRewrite(MovInstOpInterface op,
-                                  PatternRewriter &rewriter) const {
-  // CopyOpPattern handles lsir::CopyOp specifically.
-  if (isa<lsir::CopyOp>(op))
-    return failure();
-  RegisterTypeInterface dstTy =
-      dyn_cast<RegisterTypeInterface>(op.getDestOperand().getType());
-  RegisterTypeInterface srcTy =
-      dyn_cast<RegisterTypeInterface>(op.getSrcOperand().getType());
-  if (!srcTy || !dstTy)
-    return failure();
-  if (!srcTy.hasAllocatedSemantics() || !dstTy.hasAllocatedSemantics())
-    return failure();
-  if (srcTy != dstTy)
-    return rewriter.notifyMatchFailure(
-        op, "source and destination types do not match");
-  rewriter.eraseOp(op);
-  return success();
-}
-
 //===----------------------------------------------------------------------===//
 // RegisterColoring pass
 //===----------------------------------------------------------------------===//
@@ -659,11 +538,6 @@ LogicalResult RegisterColoring::run(FunctionOpInterface funcOp) {
   RewritePatternSet patterns(&getContext());
   patterns.add<InstRewritePattern, MakeRegisterRangeOpPattern,
                RegInterferenceOpPattern>(&getContext());
-  // CopyOpPattern handles lsir::CopyOp with dedicated logic. MovInstOpPattern
-  // guards against lsir::CopyOp explicitly because CopyOpPattern can return
-  // failure for it; the higher benefit ensures CopyOpPattern runs first.
-  patterns.add<CopyOpPattern>(&getContext(), /*benefit=*/2);
-  patterns.add<MovInstOpPattern>(&getContext(), /*benefit=*/1);
   FrozenRewritePatternSet frozenPatterns(std::move(patterns));
   if (failed(applyPatternsGreedily(
           funcOp, frozenPatterns,
