+++ test/Dialect/AMDGCN/Transforms/register-coloring.mlir
@@ -1,4 +1,4 @@
-// RUN: aster-opt %s --amdgcn-register-coloring --cse --split-input-file | FileCheck %s
+// RUN: aster-opt %s --amdgcn-register-coloring --amdgcn-post-reg-alloc-legalization --cse --split-input-file | FileCheck %s
 
 // CHECK-LABEL:   amdgcn.kernel @range_allocations {
 // CHECK-DAG:       %[[VAL_0:.*]] = alloca : !amdgcn.vgpr<0>
